Split model classes out of message classes
