+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/messages/scp/IPTagGetInfo.java
@@ -16,7 +16,6 @@
  */
 package uk.ac.manchester.spinnaker.messages.scp;
 
-import static java.lang.Byte.toUnsignedInt;
 import static uk.ac.manchester.spinnaker.messages.model.IPTagCommand.TTO;
 import static uk.ac.manchester.spinnaker.messages.scp.IPTagFieldDefinitions.COMMAND_FIELD;
 import static uk.ac.manchester.spinnaker.messages.scp.SCPCommand.CMD_IPTAG;
@@ -24,7 +23,6 @@ import static uk.ac.manchester.spinnaker.messages.scp.SCPCommand.CMD_IPTAG;
 import java.nio.ByteBuffer;
 
 import uk.ac.manchester.spinnaker.machine.HasChipLocation;
-import uk.ac.manchester.spinnaker.messages.model.IPTagTimeOutWaitTime;
 import uk.ac.manchester.spinnaker.messages.model.UnexpectedResponseCodeException;
 
 /**
@@ -54,32 +52,11 @@ public class IPTagGetInfo extends SCPRequest<IPTagGetInfo.Response> {
 		return new IPTagGetInfo.Response(buffer);
 	}
 
-	/** Information about a tag pool on an Ethernet-connected chip. */
-	public static final class TagInfo {
-		/**
-		 * The timeout for transient IP tags (i.e., responses to SCP commands).
-		 */
-		public final IPTagTimeOutWaitTime transientTimeout;
-
-		/** The count of the IP tag pool size. */
-		public final int poolSize;
-
-		/** The count of the number of fixed IP tag entries. */
-		public final int fixedSize;
-
-		TagInfo(ByteBuffer buffer) {
-			transientTimeout = IPTagTimeOutWaitTime.get(buffer.get());
-			buffer.get(); // skip 1 (sizeof(iptag_t) isn't relevant to us)
-			poolSize = toUnsignedInt(buffer.get());
-			fixedSize = toUnsignedInt(buffer.get());
-		}
-	}
-
 	/** An SCP response to a request for information about IP tags. */
 	protected final class Response
 			extends PayloadedResponse<TagInfo, RuntimeException> {
 		Response(ByteBuffer buffer) throws UnexpectedResponseCodeException {
-			super("Get IP Tag Info", CMD_IPTAG, buffer);
+			super("Get IP Tag Table Info", CMD_IPTAG, buffer);
 		}
 
 		@Override
