+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/messages/scp/TagInfo.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2018-2023 The University of Manchester
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package uk.ac.manchester.spinnaker.messages.scp;
+
+import static java.lang.Byte.toUnsignedInt;
+
+import java.nio.ByteBuffer;
+
+import uk.ac.manchester.spinnaker.messages.model.IPTagTimeOutWaitTime;
+
+/** Information about a tag pool on an Ethernet-connected chip. */
+public final class TagInfo {
+	/**
+	 * The timeout for transient IP tags (i.e., responses to SCP commands).
+	 */
+	public final IPTagTimeOutWaitTime transientTimeout;
+
+	/** The count of the IP tag pool size. */
+	public final int poolSize;
+
+	/** The count of the number of fixed IP tag entries. */
+	public final int fixedSize;
+
+	TagInfo(ByteBuffer buffer) {
+		transientTimeout = IPTagTimeOutWaitTime.get(buffer.get());
+		buffer.get(); // skip 1 (sizeof(iptag_t) isn't relevant to us)
+		poolSize = toUnsignedInt(buffer.get());
+		fixedSize = toUnsignedInt(buffer.get());
+	}
+}
